Refactors my forms based on PR comments
